requirements.txt
django==2.2.13
django-tastypie==0.14.2
django-crispy-forms==1.7.2
pytz
defusedxml==0.5.0
Pillow==7.1.0
sorl-thumbnail==12.6.3
libsass==0.16.0
django-compressor==2.2
django-sass-processor==0.7.2
pycodestyle
pytest
pytest-django
openpyxl==3.0.3
tablib==2.0.0
djangorestframework==3.12.1
markdown==3.3.1
django-filter==2.4.0